Data.EXT2.Info.Types

Description

Types used to expose ext2 information in a pretty way.

Documentation

type ByteAmount = Integer

data EXT2Info

Constructors

EXT2Info

Fields

totalSize :: ByteAmount

usedFileSpaceSize :: ByteAmount

unusedFileSpaceSize :: ByteAmount

spaceUsed :: ByteAmount

numInodes :: Integer

numFiles :: Integer

numDirectories :: Integer

numBlockGroups :: Integer

blockSize :: ByteAmount

data EXT2Error

Constructors

InvalidMagicNumber

InconsistentSuperblocks

InconsistentBGDT

UnreachableUsedInode

ReachableUnusedInode

DirectoryInodeUnused

UsedDataBlockWithoutInode

UnusedDataBlockWithInode

InvalidFileCount

InvalidDirectoryCount

data IntegrityStatus a

IntegrityStatus is isomorphic to Maybe, but typically Nothing is used to show a missing value or an error condition, rather than a success (and we consider being consistent to be a success value).

Constructors

Consistent

Inconsistent a

maybeIso :: Iso' (IntegrityStatus a) (Maybe a)