Data.EXT2.Superblock

Description

This module exposes functions and types for dealing with ext2 superblocks.

Documentation

data FileSystemState

Constructors

StateClean

StateErrors

data ErrorHandlingMethod

Constructors

MethodIgnore

MethodRemount

MethodPanic

data OperatingSystem

Constructors

Linux

HURD

MASIX

FreeBSD

Other

data Superblock

Constructors

Superblock

Fields

sbInodesCount :: Integer

Total number of inodes in the file system

sbBlocksCount :: Integer

Total number of blocks in the file system

sbRBlocksCount :: Integer

Number of blocks reserved for the superuser

sbFreeBlocksCount :: Integer

Number of unallocated blocks

sbFreeInodesCount :: Integer

Number of unallocated inodes

sbFirstDataBlock :: Integer

Block number of the block containing the superblock

sbLogBlockSize :: Integer

log2(block size) - 10

sbLogFragSize :: Integer

log2(fragment size) - 10

sbBlocksPerGroup :: Integer

Number of blocks in each group

sbFragsPerGroup :: Integer

Number of gragments in each group

sbInodesPerGroup :: Integer

Number of inodes in each group

sbMTime :: UnixTime

Last mount time

sbWTime :: UnixTime

Last write time

sbMntCount :: Integer

Number of mounts since last consistency check

sbMaxMntCount :: Integer

Number of allowed mounts before requiring a consistency check

sbMagic :: Integer

ext2 signature: 0xef53

sbState :: FileSystemState

Filesystem state

sbErrors :: ErrorHandlingMethod

What to do on an error condition

sbMinorRevLevel :: Integer

Minor portion of version

sbLastCheck :: UnixTime

Time of last consistency check

sbCheckInterval :: UnixTime

Interval between forced consistency checks

sbCreatorOs :: OperatingSystem

Operating system ID

sbRevLevel :: Integer

Major portion of version

sbDefResuid :: Integer

User ID that can use reserved blocks

sbDefResgid :: Integer

Group ID that can use reserved blocks

fetchSuperblock :: Handle -> IO (Either EXT2Error Superblock)

numBlockGroups :: Superblock -> Integer

blockOffset :: Superblock -> Integer -> Integer

checkIdent :: Superblock -> Either EXT2Error Superblock

Superblock Lenses

wTime :: Lens' Superblock UnixTime

state :: Lens' Superblock FileSystemState

revLevel :: Lens' Superblock Integer

rBlocksCount :: Lens' Superblock Integer

mntCount :: Lens' Superblock Integer

minorRevLevel :: Lens' Superblock Integer

maxMntCount :: Lens' Superblock Integer

magic :: Lens' Superblock Integer

mTime :: Lens' Superblock UnixTime

logFragSize :: Lens' Superblock Integer

logBlockSize :: Lens' Superblock Integer

lastCheck :: Lens' Superblock UnixTime

inodesPerGroup :: Lens' Superblock Integer

inodesCount :: Lens' Superblock Integer

freeInodesCount :: Lens' Superblock Integer

freeBlocksCount :: Lens' Superblock Integer

fragsPerGroup :: Lens' Superblock Integer

firstDataBlock :: Lens' Superblock Integer

errors :: Lens' Superblock ErrorHandlingMethod

defResuid :: Lens' Superblock Integer

defResgid :: Lens' Superblock Integer

creatorOs :: Lens' Superblock OperatingSystem

checkInterval :: Lens' Superblock UnixTime

blocksPerGroup :: Lens' Superblock Integer

blocksCount :: Lens' Superblock Integer