Data.EXT2.Info.Types

Description

Types used to expose ext2 information in a pretty way.

Documentation

type ByteAmount = Integer

data EXT2Info

Constructors

EXT2Info

Fields

ext2TotalSize :: ByteAmount

ext2UsedFileSpaceSize :: ByteAmount

ext2UnusedFileSpaceSize :: ByteAmount

ext2SpaceUsed :: ByteAmount

ext2NumInodes :: Integer

ext2NumFiles :: Integer

ext2NumDirectories :: Integer

ext2NumBlockGroups :: Integer

ext2BlockSize :: ByteAmount

ext2StateClean :: Bool

data EXT2Error

Constructors

GeneralError

InvalidMagicNumber

InconsistentSuperblocks

InconsistentBGDT

UnreachableUsedInode

ReachableUnusedInode

DirectoryInodeUnused

UsedDataBlockWithoutInode

UnusedDataBlockWithInode

InvalidFileCount

InvalidDirectoryCount

data IntegrityStatus a

IntegrityStatus is isomorphic to Maybe, but typically Nothing is used to show a missing value or an error condition, rather than a success (and we consider being consistent to be a success value).

Constructors

Consistent

Inconsistent a

maybeIso :: Iso' (IntegrityStatus a) (Maybe a)

EXT2Info lenses.

totalSize :: HasTotalSize s a => Lens' s a

usedFileSpaceSize :: HasUsedFileSpaceSize s a => Lens' s a

unusedFileSpaceSize :: HasUnusedFileSpaceSize s a => Lens' s a

spaceUsed :: HasSpaceUsed s a => Lens' s a

numInodes :: HasNumInodes s a => Lens' s a

numFiles :: HasNumFiles s a => Lens' s a

numDirectories :: HasNumDirectories s a => Lens' s a

numBlockGroups :: HasNumBlockGroups s a => Lens' s a

blockSize :: HasBlockSize s a => Lens' s a

stateClean :: HasStateClean s a => Lens' s a